Previous patch (revision 1185) for the comments-related segfault was incomplete, and as such, it broke keywords in files whose metadata had keywords but no comments.
This patch fixes that (and still fixes the case where one searches on 
comments and hits a file with keywords but no comments).  Lastly, I would 
imagine that there exists a failure mode where one searches for a keyword and 
hits a file with a comment but no keyword, and I imagine this patch should fix 
that case as well.

Patch by Omari Stephens (aka xsdg).