Add "lock" functionality to keep FileDatas in memory
Adds two primitives to keep a FileData in memory even after its refcount
drops to zero.  Intended to be used in situations when an entire group of
FileDatas _should_ stay in memory as an optimization, even if the code
would continue to function properly even if the FileData were freed.