Remove over optimization
Thanks Uwe Ohse <uwe@ohse.de> for this find.

"it looks good, but is dead slow, for each and every call to
gdk_cairo_set_source_pixbuf takes at least 15 milliseconds on my not
really slow machine, which amounts to over 2 seconds if 144 tiles have
to be drawn (tested with many jpegs and a few large pngs).

i propose to remove this optimization until someone fixes it. After
removal geeqie is fast, again."