Feature Request #70 Navigation menu option and button for parent directory
This request was made some years ago:

https://sourceforge.net/p/geeqie/feature-requests/70/

I have attached a patch which implements this. Also, keyboard-only users
will be able to use keyboard-accelerators to map a key to the Up function.

Unfortunately there is a usability problem. If the configuration file
geeqirc.xml exists (which it will except for first-time installations)
geeqie will not look at the programmed defaults, but will create the
toolbar from the config file - and the Up icon will not be displayed.

If the user first executes a command such as this:

sed -i '/toolitem action = "Back"/a <toolitem action = "Up" />'
~/.config/geeqie/geeqierc.xml

then all will be well.

This is not ideal, but I see no other solution.

Colin Clark