Allow to choose to show parent folder
With e9fc136 we learned buttons for parent folder in directory list.
This disabled the ".." handle in the view itself.

This was found to be not optimal as the users are used to it.

With this patch, that entry is now configurable in preferences. It is
enabled by default but could be disabled if the user don't like it.