Reintroduce file_data_new_simple()
This was removed by Vladimir with
efc8096527d6f743d17c44b9982616a169973f39 to always check for sidecars.

However, the newly introduced function file_data_new_group() is very
expensive. Having it in collection (in sort manager) where the dirs of
that files have huge amount of files inside, that is simply blocking
whole geeqie.

It might be that the file_data_new_group() function need to be disarmed
on other places too.