Bug Fix #233 Custom Border
https://github.com/BestImageViewer/geeqie/issues/233

In the change from v1.1 to v1.2 (GTK3 and cairo were part of this), the
custom border was ignored and simply set to black.
Perhaps it was not possible to get a "good" solution with the current
design.
Possibly this change is not the best solution, but nevertheless it
should be sufficient until a better version is produced.