Fixing Memory leak
I also suffer from what I believe is this bug. It was not present in
geeqie 1.0 and was simple for me to reproduce in 1.2+. I open a
directory containing a few thousand jpg files of around 100k each and
just keep hitting the key for "Next Image". After around a minute my
computer will freeze.

Given this reproducibility, I ran git bisect and found the commit that
introduced the behaviour for me was the following

    commit b4c4a924121c3aa6b2f1eb7570dd3f419d2f4d83
    Author: Vladimir Nadvornik <nadvornik@suse.cz>
    Date:   Thu Aug 2 00:15:16 2012 +0200

        use cairo for drawing

I then noticed what I think is a missing cairo_destroy call in
cellrenderericon.c. The following patch solves the problem completely
for me as far as I can tell.