Move exif auto-rotate option to a menu
Move the exif auto-rotate option from preferences to the
edit/orientation menu. The shortcut key is Alt-X. There is also a button
on the status bar.
The intention is to make it easy for the user to make a comparison
beteen the actual and auto-rotated view as an aid to detecting
orientations made erroneously e.g. scanned images which have been
mirrored by mistake.

Expanded documentation on rotation commands.

Preserve the setting of the rectangular select option.