Bug fix: Re-size images in GTK3
When compiled with GTK3 if an image is opened with "View in new window"
the minimum window size is determined by the image size and screen size
etc. If a file selection is opened, the minimum window size is set by
the first image displayed.
This feature is determined by a call to gtk_widget_set_size_request().

When compiled with GTK2, the call to gtk_widget_set_size_request() seems
to be ignored, and the window can be resized at will.

The GTK2 style is much more practical, so the call to
gtk_widget_set_size_request() is ignored in GTK3.