Read metadata in the idle loop
Selecting sort-by DateOriginal, DateDigitized or Rating caused Geeqie to
freeze while the metadata was read for the whole folder contents.

This commit reads those data in the idle loop and thus prevents the
freeze.
There is an option in Preferences/Metadata to autmatically read the
metadata in the idle loop whenever a new folder is selected.