Collection preview
If the folder containing collection files (usually
$HOME/.local/share/geeqie/collections) is opened, the image pane will
show a montage of the thumbnails (to a maximum of 20) of the images in
the selected collection.
Requires ImageMagick.

This is a simplistic implementation, and should be improved e.g.

Use collect-io.c to parse the collection file
Use cached thumbnails instead of re-creating
Use Geeqie to create the montage instead of using ImageMagick