Avoid non-portable == for the test command
The test command doesn't accept `==` as operator. It should be a single `=` for portable use. The `==` is a gnu extension.

Note that the x-trick hasn't been needed for a long, long time. You can reliably write it with quotes:
```
if test "$gtk3" = yes; then
```
but I left that alone since it doesn't hurt other than being ugly.