Fix #437:  Delete to standard Trash directory
https://github.com/BestImageViewer/geeqie/issues/437

The method of file deletion is changed. The Preferences/Behaviour option
"Safe delete" is removed.

The user has the option to choose Geeqie-defined trash bin or
system-defined trash bin as the normal mode - set in
Preferences/Behaviour.

Menus have two entries: Move to trash and Permanent delete.

The Delete key moves to trash; shift-Delete deletes permanently.

Both Move to trash and Permanent delete have user-selectable
confirmation, set in Preferences/Behaviour.

This means the function of the Delete key is changed from Permanent
delete to Move to trash, which is a fail-safe change.