Allow to disable trash completely
The posibilito to not use trash at all was lost with implementation of
system trash.

However, usually one does not want the system to keep megabytes of trash
lying around. That option allows to switch trash of completely.

The implementation, however, is a bit dirty, maybe someone else has a
better Idea to do it.